Extract `WordPressComRESTAPIVersion` to dedicated file
