+++ Sources/WordPressKit/WordPressKit.h
@@ -8,6 +8,7 @@ FOUNDATION_EXPORT const unsigned char WordPressKitVersionString[];
 
 #import <WordPressKit/FilePart.h>
 #import <WordPressKit/WordPressComRESTAPIInterfacing.h>
+#import <WordPressKit/WordPressComRESTAPIVersion.h>
 
 #import <WordPressKit/ServiceRemoteWordPressComREST.h>
 #import <WordPressKit/ServiceRemoteWordPressXMLRPC.h>
